pyproject.toml
[project]
name = "sue-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jsonpatch>=1.33",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.9",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "urllib3>=2.5.0",
]
